infra/scripts/requirements.txt
azure-identity
azure-kusto-data
azure-kusto-ingest
pandas
azure-eventhub